Update auto-renewal related function APIs
To be the following:
- Dnsimple.Registrar.enable_auto_renewal
- Dnsimple.Registrar.disable_auto_renewal

These functions will be under the same module as the existing registrar
related ones so we expose the same API as in other clients.
