+++ test/dnsimple/registrar_test.exs
@@ -119,6 +119,35 @@ defmodule Dnsimple.RegistrarTest do
   end
 
 
+  test ".enable_auto_renewal" do
+    fixture = "enableDomainAutoRenewal/success.http"
+    method  = "put"
+    url     = "#{@client.base_url}/v2/1010/registrar/domains/example.com/auto_renewal"
+    attributes  = []
+    {:ok, body} = Poison.encode(attributes)
+
+    use_cassette :stub, ExvcrUtils.response_fixture(fixture, method: method, url: url, request_body: body) do
+      {:ok, response} = @service.enable_auto_renewal(@client, "1010", "example.com")
+
+      assert response.__struct__ == Dnsimple.Response
+      assert is_nil(response.data)
+    end
+  end
+
+  test ".disable_auto_renewal" do
+    fixture = "disableDomainAutoRenewal/success.http"
+    method  = "delete"
+    url     = "#{@client.base_url}/v2/1010/registrar/domains/example.com/auto_renewal"
+
+    use_cassette :stub, ExvcrUtils.response_fixture(fixture, method: method, url: url) do
+      {:ok, response} = @service.disable_auto_renewal(@client, "1010", "example.com")
+
+      assert response.__struct__ == Dnsimple.Response
+      assert is_nil(response.data)
+    end
+  end
+
+
   test ".whois_privacy" do
     fixture = "getWhoisPrivacy/success.http"
     method  = "get"
